refactor: 각 예외처리 방식 통일 (#40)
