+++ api-module/src/main/java/org/devridge/api/domain/emailverification/service/EmailVerificationService.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.devridge.api.domain.emailverification.dto.request.SendEmailRequest;
 import org.devridge.api.domain.emailverification.entity.EmailVerification;
 import org.devridge.api.domain.emailverification.repository.EmailVerificationRepository;
-import org.devridge.api.exception.email.EmailVerificationInvalidException;
+import org.devridge.api.domain.emailverification.exception.EmailVerificationInvalidException;
 import org.devridge.api.util.RandomGeneratorUtil;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
@@ -60,7 +60,7 @@ public class EmailVerificationService {
     @Transactional
     public void checkVerificationCode(String email, String code) {
         EmailVerification emailVerification = emailVerificationRepository.findTopByReceiptEmailOrderByCreatedAtDesc(email)
-                .orElseThrow(() -> new EmailVerificationInvalidException());
+                .orElseThrow(() -> new EmailVerificationInvalidException(404, "해당 데이터를 찾을 수 없습니다."));
 
         LocalDateTime current = LocalDateTime.now();
