+++ api-module/src/main/java/org/devridge/api/domain/sociallogin/exception/SocialLoginException.java
@@ -1,10 +1,7 @@
-package org.devridge.api.exception.sociallogin;
+package org.devridge.api.domain.sociallogin.exception;
 
 import org.devridge.api.exception.common.BaseException;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
 
-@ResponseStatus(HttpStatus.BAD_REQUEST)
 public class SocialLoginException extends BaseException {
 
     public SocialLoginException(int code, String message) {
